Revert "Install.py: Prevent prompts from breaking apt"
This reverts commit 5995ae7e78dd19f4036f891db9db9fec97d6eab5.

With the changes to ceph-qa-chef and the teuthology kernel task,
we're no longer touching packaged file /etc/grub.d/10_linux, which
was the reason for this apt forcing.  Remove so that we find other
package problems that might be masked by this; we can always
put it back if there are such problems until we can fix those as well.

Signed-off-by: Dan Mick <dan.mick@inktank.com>