rgw tests: clean up immediately after the test
There's no need for an explicit cleanup function, so move it back
to where it came from (except in s3roundtrip, which did not have it).

Instead, since these use a nested contextmanager, pass through
and yield to the top-level run_tasks after the nested
contextmanager has finished (and thus run all the cleanup steps
in the subtasks for this test).

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>