test: unittest_rgw_meta isn't a unit test
This requires a Ceph and RGW install in order to run, and it takes
run-time parameters. This isn't a unit test and it certainly can't
run as part of make check, so move it into the debug programs group
and name it like we name our other tests which require infrastructure.

Signed-off-by: Greg Farnum <greg@inktank.com>