ceph: backward-compatibility hack: blank line before JSON output
Many JSON commands (osd dump, et. al.) used to print a status
line first before the actual output; this has been fixed, but there
are scripts/tools/etc. that expect it.  A simple compatibility hack
is to output a blank line, which won't confuse properly-written
JSON parsers, but will allow the tools-with-workarounds to continue
to work.

Signed-off-by: Dan Mick <dan.mick@inktank.com>